libraries/botbuilder-community-dialogs-prompts/requirements.txt
aiounittest>=1.3.0
botbuilder-core>=4.5.0b5
botbuilder-schema>=4.5.0b5
botbuilder-dialogs>=4.5.0b5
recognizers_text_suite>=1.0.2a2